v2: package files re-organization
This change splits large files from the `v2` sub-package into smaller
files, usually one per API resource.
